TS-8032 Restructure and add new base class
