+++ teamscale-maven-plugin/src/main/java/com/teamscale/maven/tia/TiaMojoBase.java
@@ -1,11 +1,11 @@
-package com.teamscale.tia.maven;
+package com.teamscale.maven.tia;
 
+import com.teamscale.maven.TeamscaleMojoBase;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.model.Plugin;
-import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.codehaus.plexus.util.xml.Xpp3Dom;
@@ -52,42 +52,7 @@ import java.util.Map;
  * <p>
  * The log file of the agent is written to {@code ${project.build.directory}/tia/agent.log}.
  */
-public abstract class TiaMojoBase extends AbstractMojo {
-
-	/**
-	 * The URL of the Teamscale instance to which the recorded coverage will be uploaded.
-	 */
-	@Parameter(required = true)
-	public String teamscaleUrl;
-
-	/**
-	 * The Teamscale project to which the recorded coverage will be uploaded
-	 */
-	@Parameter(required = true)
-	public String projectId;
-
-	/**
-	 * The username to use to perform the upload. Must have the "Upload external data" permission for the {@link
-	 * #projectId}. Can also be specified via the Maven property {@code teamscale.username}.
-	 */
-	@Parameter(property = "teamscale.username", required = true)
-	public String username;
-
-	/**
-	 * Teamscale access token of the {@link #username}. Can also be specified via the Maven property {@code
-	 * teamscale.accessToken}.
-	 */
-	@Parameter(property = "teamscale.accessToken", required = true)
-	public String accessToken;
-
-	/**
-	 * You can optionally use this property to override the code commit to which the coverage will be uploaded. Format:
-	 * {@code BRANCH:UNIX_EPOCH_TIMESTAMP_IN_MILLISECONDS}
-	 * <p>
-	 * If no end commit is manually specified, the plugin will try to determine the currently checked out Git commit.
-	 */
-	@Parameter
-	public String endCommit;
+public abstract class TiaMojoBase extends TeamscaleMojoBase {
 
 	/**
 	 * You can optionally specify which code should be included in the coverage instrumentation. Each pattern is applied
@@ -131,36 +96,11 @@ public abstract class TiaMojoBase extends AbstractMojo {
 	@Parameter
 	public String[] additionalAgentOptions;
 
-
 	/**
 	 * Changes the log level of the agent to DEBUG.
 	 */
 	@Parameter(defaultValue = "false")
 	public boolean debugLogging;
-
-	/**
-	 * Whether to skip the execution of this Mojo.
-	 */
-	@Parameter(defaultValue = "false")
-	public boolean skip;
-
-	/**
-	 * Map of resolved Maven artifacts. Provided automatically by Maven.
-	 */
-	@Parameter(property = "plugin.artifactMap", required = true, readonly = true)
-	public Map<String, Artifact> pluginArtifactMap;
-
-	/**
-	 * The project build directory (usually: {@code ./target}). Provided automatically by Maven.
-	 */
-	@Parameter(defaultValue = "${project.build.directory}")
-	public String projectBuildDir;
-
-	/**
-	 * The running Maven session. Provided automatically by Maven.
-	 */
-	@Parameter(defaultValue = "${session}")
-	public MavenSession session;
 
 	private Path targetDirectory;
 	private String resolvedEndCommit;
@@ -324,21 +264,6 @@ public abstract class TiaMojoBase extends AbstractMojo {
 	private Path findAgentJarFile() {
 		Artifact agentArtifact = pluginArtifactMap.get("com.teamscale:teamscale-jacoco-agent");
 		return agentArtifact.getFile().toPath();
-	}
-
-	private String resolveEndCommit() throws MojoFailureException {
-		if (StringUtils.isNotBlank(endCommit)) {
-			return endCommit;
-		}
-
-		Path basedir = session.getCurrentProject().getBasedir().toPath();
-		try {
-			GitCommit commit = GitCommit.getGitHeadCommitDescriptor(basedir);
-			return commit.branch + ":" + commit.timestamp;
-		} catch (IOException e) {
-			throw new MojoFailureException("You did not configure an <endCommit> in the pom.xml" +
-					" and I could also not determine the checked out commit in " + basedir + " from Git", e);
-		}
 	}
 
 	/**
