+++ teamscale-maven-plugin/src/main/resources/com/teamscale/maven/tia/logback-agent.xml
